6-size.c

+++ 0x00-hello_world/6-size.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 /**
- * main - the size of various types on the computer
+ * main - the size of various types 
  * Return: Always 0 (Success)
  */
 int main(void)
@@ -10,10 +10,10 @@ int main(void)
 	long int c;
 	long long int d;
 	float f;
-printf("Size of a char: %lu byte(s), sizeof(a));
-printf("Size of a int: %lu byte(s)\n", sizeof(b));
-printf("Size of a long int: %lu bytes(s)\n", sizeof(c));
-printf("Size of a long long int: %lu bytes(s)\n", sizeof(d));
-printf("Size of a float: %lu bytes(s)\n", sizeof(f));
+printf("Size of a char: %lu byte(s), (unsigned long)sizeof(a));
+printf("Size of a int: %lu byte(s)\n", (unsigned long)sizeof(b));
+printf("Size of a long int: %lu bytes(s)\n", (unsigned long)sizeof(c));
+printf("Size of a long long int: %lu bytes(s)\n", (unsigned long)sizeof(d));
+printf("Size of a float: %lu bytes(s)\n", (unsigned long)sizeof(f));
 	return (0);
 }
